# tools/old_gateway/docs/swagger.yaml
definitions:
  dto.AdminLoginInput:
    properties:
      password:
        description: 密码
        example: passwd
        type: string
      username:
        description: admin用户名
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  dto.ChangePwdInput:
    properties:
      password:
        description: 密码
        example: changepasswd
        type: string
    required:
    - password
    type: object
  dto.ServiceAddHttpInput:
    properties:
      black_list:
        description: 黑名单ip
        type: string
      clientip_flow_limit:
        description: 客户端ip限流
        type: integer
      header_transfor:
        description: header转换
        type: string
      ip_list:
        description: ip列表
        type: string
      need_https:
        description: 支持https
        type: integer
      need_strip_uri:
        description: 启用strip_uri,注意这个功能和url重写冲突性
        type: integer
      need_websocket:
        description: 是否支持websocket
        type: integer
      open_auth:
        description: service_access_control
        type: integer
      round_type:
        description: service_load_balance
        type: integer
      rule:
        description: 域名或者前缀(路径/add)
        type: string
      rule_type:
        description: service_http_rule
        type: integer
      service_desc:
        description: 服务描述
        type: string
      service_flow_limit:
        description: 服务端限流
        type: integer
      service_name:
        description: service_info
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        type: integer
      url_rewrite:
        description: url重写功能
        type: string
      weight_list:
        description: "\b权重列表"
        type: string
      white_list:
        description: 白名单ip
        type: string
    required:
    - ip_list
    - rule
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceAddTcpInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      header_transfor:
        type: string
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceUpdateHTTPInput:
    properties:
      black_list:
        description: 黑名单ip
        type: string
      clientip_flow_limit:
        description: "\b客户端ip限流"
        type: integer
      header_transfor:
        description: header转换
        type: string
      id:
        description: 服务ID
        example: 62
        type: integer
      ip_list:
        description: ip列表
        example: 127.0.0.1:80
        type: string
      need_https:
        description: 支持https
        type: integer
      need_strip_uri:
        description: 启用strip_uri
        type: integer
      need_websocket:
        description: 是否支持websocket
        type: integer
      open_auth:
        description: 关键词
        type: integer
      round_type:
        description: 轮询方式
        type: integer
      rule:
        description: 域名或者前缀
        example: /test_http_service_indb
        type: string
      rule_type:
        description: 接入类型
        type: integer
      service_desc:
        description: 服务描述
        example: test_http_service_indb
        type: string
      service_flow_limit:
        description: 服务端限流
        type: integer
      service_name:
        description: 服务名
        example: test_http_service_indb
        type: string
      upstream_connect_timeout:
        description: 建立连接超时, 单位s
        type: integer
      upstream_header_timeout:
        description: 获取header超时, 单位s
        type: integer
      upstream_idle_timeout:
        description: 链接最大空闲时间, 单位s
        type: integer
      upstream_max_idle:
        description: 最大空闲链接数
        type: integer
      url_rewrite:
        description: url重写功能
        type: string
      weight_list:
        description: "\b权重列表"
        example: "50"
        type: string
      white_list:
        description: 白名单ip
        type: string
    required:
    - id
    - ip_list
    - rule
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.ServiceUpdateTcpInput:
    properties:
      black_list:
        type: string
      clientip_flow_limit:
        type: integer
      forbid_list:
        type: string
      id:
        type: integer
      ip_list:
        type: string
      open_auth:
        type: integer
      port:
        type: integer
      round_type:
        type: integer
      service_desc:
        type: string
      service_flow_limit:
        type: integer
      service_name:
        type: string
      weight_list:
        type: string
      white_host_name:
        type: string
      white_list:
        type: string
    required:
    - id
    - ip_list
    - port
    - service_desc
    - service_name
    - weight_list
    type: object
  dto.TenantAddHttpInput:
    properties:
      app_id:
        type: string
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - app_id
    - name
    type: object
  dto.TenantUpdateInput:
    properties:
      app_id:
        type: string
      id:
        type: integer
      name:
        type: string
      qpd:
        type: integer
      qps:
        type: integer
      secret:
        type: string
      white_ips:
        type: string
    required:
    - id
    - name
    - secret
    type: object
  middleware.Response:
    properties:
      data:
        type: object
      errmsg:
        description: 响应消息
        type: string
      errno:
        description: 响应码
        type: integer
      stack:
        description: 堆栈信息
        type: object
      trace_id:
        description: 追踪id
        type: object
    type: object
info:
  contact: {}
  license: {}
paths:
  /admin/admin_info:
    get:
      consumes:
      - application/json
      description: admin信息
      operationId: /admin/admin_info
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: admin信息
      tags:
      - admin接口
  /admin/change_pwd:
    post:
      consumes:
      - application/json
      description: 修改密码
      operationId: /admin/change_pwd
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePwdInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 修改密码
      tags:
      - admin接口
  /admin_login/login:
    post:
      consumes:
      - application/json
      description: 登录
      operationId: /admin_login/login
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AdminLoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 登录
      tags:
      - admin接口
  /admin_login/logout:
    get:
      consumes:
      - application/json
      description: 退出
      operationId: /admin_login/logout
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 退出
      tags:
      - admin接口
  /market/flow_stat:
    get:
      consumes:
      - application/json
      description: 流量统计
      operationId: /market/flow_stat
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 流量统计
      tags:
      - market
  /market/panel:
    get:
      consumes:
      - application/json
      description: 面板数据统计
      operationId: /market/panel
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 面板数据统计
      tags:
      - market
  /market/service_stat:
    get:
      consumes:
      - application/json
      description: service统计
      operationId: /market/service_stat
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: service按类型统计
      tags:
      - market
  /service/add_http:
    post:
      consumes:
      - application/json
      description: 添加HTTP服务
      operationId: /service/add_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddHttpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 添加HTTP服务
      tags:
      - service管理
  /service/add_tcp:
    post:
      consumes:
      - application/json
      description: 添加tcp服务
      operationId: /service/add_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceAddTcpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 添加tcp服务
      tags:
      - service管理
  /service/delete:
    get:
      consumes:
      - application/json
      description: service删除
      operationId: /service/delete
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: service删除
      tags:
      - service管理
  /service/detail:
    get:
      consumes:
      - application/json
      description: service详情
      operationId: /service/detail
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: service详情
      tags:
      - service管理
  /service/list:
    get:
      consumes:
      - application/json
      description: service列表
      operationId: /service/list
      parameters:
      - description: 关键词
        in: query
        name: info
        type: string
      - description: 每页个数
        in: query
        name: page_size
        required: true
        type: integer
      - description: 当前页数
        in: query
        name: page_num
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: service列表
      tags:
      - service管理
  /service/stat:
    get:
      consumes:
      - application/json
      description: service统计
      operationId: /service/stat
      parameters:
      - description: 服务ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: service统计(流量)
      tags:
      - service管理
  /service/update_http:
    post:
      consumes:
      - application/json
      description: 修改HTTP服务
      operationId: /service/update_http
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateHTTPInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 修改HTTP服务
      tags:
      - service管理
  /service/update_tcp:
    post:
      consumes:
      - application/json
      description: 更新tcp服务
      operationId: /service/update_tcp
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ServiceUpdateTcpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 更新tcp服务
      tags:
      - service管理
  /tenant/add:
    post:
      consumes:
      - application/json
      description: 网关租户添加
      operationId: /tenant/add
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TenantAddHttpInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户添加
      tags:
      - tenant管理
  /tenant/delete:
    get:
      consumes:
      - application/json
      description: 网关租户删除
      operationId: /tenant/delete
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户删除
      tags:
      - tenant管理
  /tenant/detail:
    get:
      consumes:
      - application/json
      description: 网关租户详情
      operationId: /tenant/detail
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户详情
      tags:
      - tenant管理
  /tenant/list:
    get:
      consumes:
      - application/json
      description: 网关租户列表
      operationId: /tenant/list
      parameters:
      - description: 搜索关键词
        in: query
        name: info
        type: string
      - description: 每页数目
        in: query
        name: page_size
        required: true
        type: string
      - description: 页码
        in: query
        name: page_num
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户列表
      tags:
      - tenant管理
  /tenant/stat:
    get:
      consumes:
      - application/json
      description: 网关租户统计
      operationId: /tenant/stat
      parameters:
      - description: 租户ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户统计
      tags:
      - tenant管理
  /tenant/update:
    post:
      consumes:
      - application/json
      description: 网关租户更新
      operationId: /tenant/update
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.TenantUpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            $ref: '#/definitions/middleware.Response'
      summary: 网关租户更新
      tags:
      - tenant管理
swagger: "2.0"